requirements.txt
python-telegram-bot==22.0
instagrapi==2.1.3
schedule==1.2.2
pytesseract==0.3.13
Pillow==11.1.0
python-json-logger==2.0.7
colorlog==6.7.0
python-dotenv==1.0.0
keyring==24.2.0
opencv-python==4.9.0.80
numpy==1.26.4
windows-curses; platform_system=="Windows"
pyreadline3; platform_system=="Windows"
colorama==0.4.6
psutil==5.9.8
